interface修改

+++ project/server/share/vector/Vector2.cc
@@ -7,7 +7,7 @@ Vector2::Vector2()
     m_y(0)
 {}
 
-Vector2::Vector2(int x, int y)
+Vector2::Vector2(float x, float y)
     :m_x(x),
     m_y(y)
 {}
@@ -18,22 +18,22 @@ Vector2::~Vector2()
     m_y = 0;
 }
 
-int Vector2::GetX(int x)
+float Vector2::GetX(float x)
 {
     return m_x;
 }
 
-int Vector2::GetY(int y)
+float Vector2::GetY(float y)
 {
     return m_y;
 }
 
-void Vector2::SetX(int x)
+void Vector2::SetX(float x)
 {
     m_x = x;
 }
 
-void Vector2::SetY(int y)
+void Vector2::SetY(float y)
 {
     m_y = y;
 }
